requirements.txt
face_recognition==1.3.0
Pillow==10.0.1
Pillow==10.0.1
